+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * You can now define the uid a node site runs as
 * Fix `Site#matches` so it also searches through defined `hostname`s
 * `Site#getAddress` will callback with an error when it fails to start
+* The callback provided to `Site#getAddress` will now accept error objects
 
 ## 0.1.0